+++ app/test/e2e/queue-status-index-deleted.spec.js
@@ -120,6 +120,60 @@ describe('STATUS_INDEX_DELETED handling process', () => {
         });
     });
 
+    it('Consume a STATUS_INDEX_DELETED message for a TASK_OVERWRITE should set the dataset status to SAVED (happy case)', async () => {
+        const fakeTask1 = await new Task(createTask(appConstants.TASK_STATUS.INIT, task.MESSAGE_TYPES.TASK_OVERWRITE)).save();
+
+        const message = {
+            id: 'e492cef7-e287-4bd8-9128-f034a3b531ef',
+            type: 'STATUS_INDEX_DELETED',
+            taskId: fakeTask1.id,
+            lastCheckedDate: '2019-03-29T08:43:08.091Z'
+        };
+
+        nock(process.env.CT_URL)
+            .patch(`/v1/dataset/${fakeTask1.datasetId}`, { status: 1 })
+            .once()
+            .reply(200, {});
+
+        const preStatusQueueStatus = await channel.assertQueue(config.get('queues.status'));
+        preStatusQueueStatus.messageCount.should.equal(0);
+        const existingTaskList = await Task.find({}).exec();
+        existingTaskList.should.be.an('array').and.have.lengthOf(1);
+
+        await channel.sendToQueue(config.get('queues.status'), Buffer.from(JSON.stringify(message)));
+
+        // Give the code a few seconds to do its thing
+        await new Promise(resolve => setTimeout(resolve, 5000));
+
+        const postQueueStatus = await channel.assertQueue(config.get('queues.status'));
+        postQueueStatus.messageCount.should.equal(0);
+
+        const createdTasks = await Task.find({}).exec();
+
+        createdTasks.should.be.an('array').and.have.lengthOf(1);
+        const createdTask = createdTasks[0];
+        createdTask.should.have.property('status').and.equal(appConstants.TASK_STATUS.SAVED);
+        createdTask.should.have.property('reads').and.equal(0);
+        createdTask.should.have.property('writes').and.equal(0);
+        createdTask.should.have.property('_id').and.equal(fakeTask1.id);
+        createdTask.should.have.property('type').and.equal(task.MESSAGE_TYPES.TASK_OVERWRITE);
+        createdTask.should.have.property('message').and.be.an('object');
+        createdTask.should.have.property('datasetId').and.equal(fakeTask1.datasetId);
+        createdTask.should.have.property('createdAt').and.be.a('date');
+        createdTask.should.have.property('updatedAt').and.be.a('date');
+        createdTask.should.have.property('logs').and.be.an('array').and.have.lengthOf(1);
+
+        const log = createdTask.logs[0];
+
+        log.should.have.property('id').and.equal(message.id);
+        log.should.have.property('taskId').and.equal(message.taskId);
+        log.should.have.property('type').and.equal(message.type);
+
+        process.on('unhandledRejection', (error) => {
+            should.fail(error);
+        });
+    });
+
     afterEach(async () => {
         await Task.deleteMany({}).exec();